% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{ColBarplot}
\alias{ColBarplot}
\title{ColBarplot}
\usage{
ColBarplot(data, d.x, d.y, lab.x, lab.y, d.title, d.source, color.pal1,
  color.pal2)
}
\arguments{
\item{data}{data to plot}

\item{d.x}{values of the data}

\item{d.y}{names of the barplot}

\item{lab.x}{label of x}

\item{lab.y}{label of y}

\item{d.title}{title of the barplot}

\item{d.source}{source of the data}

\item{color.pal1}{coulour palette}

\item{color.pal2}{coulour palette}
}
\value{
null
}
\description{
colourful barplots
}
